make refreshPage() work correctly for boolean values
Within refreshPage() incoming value is analyzed for boolean elements to be either true or false. The value was then set to "true" or "false" strings instead of boolean values.
This led to incorrect display of boolean values (e.g. 'Pause'), which have been assumed to be true, although value was string "false".